Added: The backend can now act as a very simple web server
Usually you need a web server when accessing a backend from the browser as the domain name (host) in the url must be the same for the web page and the backend (Cross-domain restrictions). Now you can just load the application from ’localhost:1337/index-debug.html’ in the browser. The backend requests will be allowed as they share the same domain name (host) in the url.

To start the backend as a web server place the model file inside the Cappuccino project, either in the root or in the Resource directory. Point out the project root directory when starting the backend.